api-stock360/services/warehouses-service/warehouses_app/security.py
import os

from jose import JWTError, jwt

SECRET_KEY = os.getenv("SECRET_KEY")
ALGORITHM = os.getenv("ALGORITHM")


def decode_token(token: str):
    try:
        payload = jwt.decode(
            token, SECRET_KEY, algorithms=[ALGORITHM], options={"verify_exp": True}
        )
        return payload
    except JWTError as e:
        print("JWT decode error:", e)
        return None
