api-stock360/services/warehouses-service/requirements.txt
fastapi
uvicorn
motor
python-jose[cryptography]
pydantic[email]
passlib[bcrypt]
prometheus-client
httpx
pytest
pytest-asyncio